Authentication feature update

+++ src/main/kotlin/com/spring/starter/model/user/LoginResponse.kt
@@ -2,6 +2,6 @@ package com.spring.starter.model.user
 
 data class LoginResponse(
     val tokenType: String,
-    val expiresInt: Int,
+    val expiresInt: Long,
     val accessToken: String,
 )
